refactor: move Draggable class file to shared directory
